feat: Register system
The system that control the users that can
read or write or not from a queue is
implemented with some metadata from the
queues, they work by setting the data into
memory but persisting the data into a file.

It's still necessary to make a queue list
become persisted in some routine of the
software because the basics are implemented
but not everything is automated.

Another problem present is that some messages
aren't being sent to the subscribers, they
just go away.