fix: Problem with message lost
There was a problem that when we wanted
to right the data into the filesystem so
we could persist the messages sent by the
clients in a queue, we sometimes overwritten
those files and lost some messages.

We implemented a system using mutexes to make
it all works just fine, at least we think it
wil work just fine.